Updated CAN ID of motor code uses (22 => 8). Fixed type problem in LeftMotorTwoTurnsCmd.

+++ practice/JanPracticeBoard/src/main/include/commands/LeftMotorTwoTurnsCmd.h
@@ -33,7 +33,7 @@ class LeftMotorTwoTurnsCmd
   bool IsFinished() override;
 
 private:
-  const int kTicksPerTurn = 0.0257125;
+  const double kTicksPerTurn = 0.0257125;
   const double kInitialPower = 0.05;
   
   DrivetrainSub *m_drivetrainSubPtr;
